requirements.txt
lightning>=2.2.0
torch>=2.1.0
torchvision>=0.16.0
pandas>=2.0.0
pyyaml>=6.0.0
loguru>=0.7.0
numpy>=1.24.0
